Robot that follows lines and fetches a load
The PikkerBot is an embedded robot project of two collage students. The
robot is designed and build by Professor Paillet Eric of the University
of Antwerp. The PikkerBot is equiped with Lego Mindstorm actuators and
sensores. These are connected with custom PCBs to a Raspberri Pi for
processing.

The PikkerBot, when started, will wait for the start and end points that
will be transmitted by a SeekerBot.

When he received those points, the bot will go and get a package at the
end point and transport it back to the start point.

The project is divided in small assignments to create the drivers for
the different components of the robot.